+++ probeye/definition/parameter.py
@@ -26,7 +26,7 @@ class Parameters(dict):
     def add_parameter(
         self,
         prm_name: str,
-        prm_type: str,
+        prm_type: str = "not defined",
         dim: Optional[int] = 1,
         domain: str = "(-oo, +oo)",
         const: Union[int, float, np.ndarray, None] = None,
@@ -612,10 +612,11 @@ class ParameterProperties:
                 f"int or None, but found {type(self._dim)}."
             )
 
-        if self._type not in ["model", "prior", "likelihood"]:
+        if self._type not in ["model", "prior", "likelihood", "not defined"]:
             raise RuntimeError(
                 f"Found invalid ParameterProperties._type attribute! It can only "
-                f"assume the three values 'model', 'prior' or 'likelihood' but found "
+                f"assume the three values 'model', 'prior', 'likelihood' (or "
+                f"'not defined' when automatic type-assignment is used) but found "
                 f"'{self._type}'."
             )